requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.0
asyncpg>=0.27.0
fastapi>=0.95.0
uvicorn>=0.22.0
stripe>=5.0.0
pytest>=7.0.0
httpx>=0.24.0
python-dateutil>=2.8.2
